Support LOOPLESS tasks
Loopless tasks are tasks that set LOOPLESS = True in their class
definition.  This causes initTask(), start(), and stop() to still be
called, but no worker thread is spun up targetting _runloop().

This allows you to implement Tasks that are dependant on other tasks,
can have inputs and methods, but do not occur based on any particular
periodic or queue-based event