Add a new build rule, "submodule_copy"
I've replaced the symlinks to externals with actual copies of the files.
The files are kept up to date automatically by developers who have the
full development environment configured (with git submodules checked out)

The submodule_copy rule takes a dictionary mapping paths in externals
to paths where the copy should reside (and stay up to date) in the
primary source.