Fix SUPER nasty bug with action='store_true' options...
Wow, how did I miss this?!  This was a pain to debug.  Logging the
option as %s was returning "False", but when doing actual logical
comparisons, it was evaluating to True.

Oops, yep, that's what happens when you have type=str with a bool.
bool('False') == True.  This isn't PHP, guys!