dbus: pass service start() and stop() handling through the main loop
The main loop used by dbus connection is spinning in separate thread in
which the DBusMainLoopTask is running. DBus bindings are not
thread-safe, and the loop has already started (DBusMainLoopTask is a
prerequisite for DBusServiceTask). The only safe way to make dbus calls
is by passing them through the thead in which the loop is being run.

Note that in typical dbus applications most of dbus calls is done before
the main loop starts or the calls are performed within the callbacks in
the main loop. The only way to reproduce this is to use idle callback
that can be added in a thread-safe manner.